+++ lib/actions.c
@@ -44,7 +44,7 @@ int mosquitto_publish_v5(struct mosquitto *mosq, int *mid, const char *topic, in
 	mosquitto_property local_property;
 	bool have_topic_alias;
 	int rc;
-	int tlen = 0;
+	size_t tlen = 0;
 	uint32_t remaining_length;
 
 	if(!mosq || qos<0 || qos>2) return MOSQ_ERR_INVAL;
@@ -89,7 +89,7 @@ int mosquitto_publish_v5(struct mosquitto *mosq, int *mid, const char *topic, in
 		}
 	}else{
 		tlen = strlen(topic);
-		if(mosquitto_validate_utf8(topic, tlen)) return MOSQ_ERR_MALFORMED_UTF8;
+		if(mosquitto_validate_utf8(topic, (int)tlen)) return MOSQ_ERR_MALFORMED_UTF8;
 		if(payloadlen < 0 || payloadlen > MQTT_MAX_PAYLOAD) return MOSQ_ERR_PAYLOAD_SIZE;
 		if(mosquitto_pub_topic_check(topic) != MOSQ_ERR_SUCCESS){
 			return MOSQ_ERR_INVAL;
@@ -97,7 +97,7 @@ int mosquitto_publish_v5(struct mosquitto *mosq, int *mid, const char *topic, in
 	}
 
 	if(mosq->maximum_packet_size > 0){
-		remaining_length = 1 + 2+tlen + payloadlen + property__get_length_all(outgoing_properties);
+		remaining_length = 1 + 2+(uint32_t)tlen + (uint32_t)payloadlen + property__get_length_all(outgoing_properties);
 		if(qos > 0){
 			remaining_length++;
 		}
@@ -112,7 +112,7 @@ int mosquitto_publish_v5(struct mosquitto *mosq, int *mid, const char *topic, in
 	}
 
 	if(qos == 0){
-		return send__publish(mosq, local_mid, topic, payloadlen, payload, qos, retain, false, outgoing_properties, NULL, 0);
+		return send__publish(mosq, local_mid, topic, (uint32_t)payloadlen, payload, (uint8_t)qos, retain, false, outgoing_properties, NULL, 0);
 	}else{
 		if(outgoing_properties){
 			rc = mosquitto_property_copy_all(&properties_copy, outgoing_properties);
@@ -137,18 +137,18 @@ int mosquitto_publish_v5(struct mosquitto *mosq, int *mid, const char *topic, in
 		}
 		if(payloadlen){
 			message->msg.payloadlen = payloadlen;
-			message->msg.payload = mosquitto__malloc(payloadlen*sizeof(uint8_t));
+			message->msg.payload = mosquitto__malloc((unsigned int)payloadlen*sizeof(uint8_t));
 			if(!message->msg.payload){
 				message__cleanup(&message);
 				mosquitto_property_free_all(&properties_copy);
 				return MOSQ_ERR_NOMEM;
 			}
-			memcpy(message->msg.payload, payload, payloadlen*sizeof(uint8_t));
+			memcpy(message->msg.payload, payload, (uint32_t)payloadlen*sizeof(uint8_t));
 		}else{
 			message->msg.payloadlen = 0;
 			message->msg.payload = NULL;
 		}
-		message->msg.qos = qos;
+		message->msg.qos = (uint8_t)qos;
 		message->msg.retain = retain;
 		message->dup = false;
 		message->properties = properties_copy;
@@ -204,9 +204,9 @@ int mosquitto_subscribe_multiple(struct mosquitto *mosq, int *mid, int sub_count
 
 	for(i=0; i<sub_count; i++){
 		if(mosquitto_sub_topic_check(sub[i])) return MOSQ_ERR_INVAL;
-		slen = strlen(sub[i]);
+		slen = (int)strlen(sub[i]);
 		if(mosquitto_validate_utf8(sub[i], slen)) return MOSQ_ERR_MALFORMED_UTF8;
-		remaining_length += 2+slen + 1;
+		remaining_length += 2+(uint32_t)slen + 1;
 	}
 
 	if(mosq->maximum_packet_size > 0){
@@ -261,13 +261,13 @@ int mosquitto_unsubscribe_multiple(struct mosquitto *mosq, int *mid, int sub_cou
 
 	for(i=0; i<sub_count; i++){
 		if(mosquitto_sub_topic_check(sub[i])) return MOSQ_ERR_INVAL;
-		slen = strlen(sub[i]);
+		slen = (int)strlen(sub[i]);
 		if(mosquitto_validate_utf8(sub[i], slen)) return MOSQ_ERR_MALFORMED_UTF8;
-		remaining_length += 2+slen;
+		remaining_length += 2U + (uint32_t)slen;
 	}
 
 	if(mosq->maximum_packet_size > 0){
-		remaining_length += 2 + property__get_length_all(outgoing_properties);
+		remaining_length += 2U + property__get_length_all(outgoing_properties);
 		if(packet__check_oversize(mosq, remaining_length)){
 			return MOSQ_ERR_OVERSIZE_PACKET;
 		}
